third_party/WebKit/LayoutTests/fast/dynamic/last-child-descendant-expected.txt

Evaluation of :last-child affecting descendant.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(b).color is green
PASS successfullyParsed is true

TEST COMPLETE

This text should be green
